moved junit uiTests of webtester-examples and webtester-example-app to new packages
