+++ webtester-examples/src/main/java/webtester/showcases/core/support/junit/Experiments/DifferentBrowserFieldModifiersExperiment.java
@@ -1,4 +1,4 @@
-package webtester.showcases.core.support.junit;
+package webtester.showcases.core.support.junit.Experiments;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
